+++ src/pools/worker.ts
@@ -37,6 +37,10 @@ export type ExitHandler<Worker extends IWorker> = (
  */
 export interface Task<Data = unknown> {
   /**
+   * Task name.
+   */
+  readonly name?: string
+  /**
    * Task input data that will be passed to the worker.
    */
   readonly data?: Data